fast_api/pyproject.toml
[project]
name = "fast-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi-mcp (>=0.2.0,<0.3.0)",
    "google-cloud-bigquery (>=3.31.0,<4.0.0)"
]

[tool.poetry]
packages = [{include = "fast_api", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
